feat: Update SideMenu and HeaderBar components
The code changes include updating the SideMenu and HeaderBar components. In the SideMenu component, the import statement for 'flutter/widgets.dart' was added. In the HeaderBar component, the toolbar height was set to 85.0 and the icon theme color was set to white. These updates enhance the app's user interface and functionality.