feat: Add IClientHttp interface and ClientHttpService implementation
This commit adds the `IClientHttp` interface and implements it with the `ClientHttpService` class. The `IClientHttp` interface defines the methods `addToken` and `get` for adding authentication tokens and making HTTP GET requests, respectively. The `ClientHttpService` class uses the `dio` package to handle HTTP requests and implements the methods defined in the interface.